requirements.txt
UnityPy>=1.9.24
requests>=2.27.1
Pillow>=9.4.0
rich>=10.2.2
requests>=2.25.0
bs4>=0.0.1
tqdm>=0.4.4
pycryptodome>=3.15.0
fuzzywuzzy>=0.18.0
Levenshtein>=0.25.0